separate namespace of settings
