+++ src/Inspector/Daemon/Reader/Context/PhpReaderContextInterface.php
@@ -16,9 +16,9 @@ namespace PhpProfiler\Inspector\Daemon\Reader\Context;
 use Amp\Promise;
 use PhpProfiler\Inspector\Daemon\Dispatcher\Message\DetachWorkerMessage;
 use PhpProfiler\Inspector\Daemon\Dispatcher\Message\TraceMessage;
-use PhpProfiler\Inspector\Settings\GetTraceSettings;
-use PhpProfiler\Inspector\Settings\TargetPhpSettings;
-use PhpProfiler\Inspector\Settings\TraceLoopSettings;
+use PhpProfiler\Inspector\Settings\GetTraceSettings\GetTraceSettings;
+use PhpProfiler\Inspector\Settings\TargetPhpSettings\TargetPhpSettings;
+use PhpProfiler\Inspector\Settings\TraceLoopSettings\TraceLoopSettings;
 
 interface PhpReaderContextInterface
 {
