+++ src/Inspector/Daemon/Reader/PhpReaderTaskInterface.php
@@ -15,10 +15,10 @@ namespace PhpProfiler\Inspector\Daemon\Reader;
 
 use Amp\Parallel\Sync\Channel;
 use Generator;
-use PhpProfiler\Inspector\Settings\GetTraceSettings;
-use PhpProfiler\Inspector\Settings\TargetPhpSettings;
-use PhpProfiler\Inspector\Settings\TargetProcessSettings;
-use PhpProfiler\Inspector\Settings\TraceLoopSettings;
+use PhpProfiler\Inspector\Settings\GetTraceSettings\GetTraceSettings;
+use PhpProfiler\Inspector\Settings\TargetPhpSettings\TargetPhpSettings;
+use PhpProfiler\Inspector\Settings\TargetProcessSettings\TargetProcessSettings;
+use PhpProfiler\Inspector\Settings\TraceLoopSettings\TraceLoopSettings;
 
 interface PhpReaderTaskInterface
 {
