+++ src/Inspector/Settings/TraceLoopSettings/TraceLoopSettings.php
@@ -11,8 +11,9 @@
 
 declare(strict_types=1);
 
-namespace PhpProfiler\Inspector\Settings;
+namespace PhpProfiler\Inspector\Settings\TraceLoopSettings;
 
+use PhpProfiler\Inspector\Settings\InspectorSettingsException;
 use Symfony\Component\Console\Input\InputInterface;
 
 final class TraceLoopSettings
