+++ tests/Inspector/Daemon/Dispatcher/DispatchTableTest.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace PhpProfiler\Inspector\Daemon\Dispatcher;
 
 use PhpProfiler\Inspector\Daemon\Reader\Context\PhpReaderContextInterface;
-use PhpProfiler\Inspector\Settings\GetTraceSettings;
-use PhpProfiler\Inspector\Settings\TargetPhpSettings;
-use PhpProfiler\Inspector\Settings\TraceLoopSettings;
+use PhpProfiler\Inspector\Settings\GetTraceSettings\GetTraceSettings;
+use PhpProfiler\Inspector\Settings\TargetPhpSettings\TargetPhpSettings;
+use PhpProfiler\Inspector\Settings\TraceLoopSettings\TraceLoopSettings;
 use PHPUnit\Framework\TestCase;
 
 class DispatchTableTest extends TestCase
