+++ tests/Inspector/Daemon/Dispatcher/WorkerPoolTest.php
@@ -17,9 +17,9 @@ use Amp\Success;
 use Mockery;
 use PhpProfiler\Inspector\Daemon\Reader\Context\PhpReaderContextCreatorInterface;
 use PhpProfiler\Inspector\Daemon\Reader\Context\PhpReaderContextInterface;
-use PhpProfiler\Inspector\Settings\GetTraceSettings;
-use PhpProfiler\Inspector\Settings\TargetPhpSettings;
-use PhpProfiler\Inspector\Settings\TraceLoopSettings;
+use PhpProfiler\Inspector\Settings\GetTraceSettings\GetTraceSettings;
+use PhpProfiler\Inspector\Settings\TargetPhpSettings\TargetPhpSettings;
+use PhpProfiler\Inspector\Settings\TraceLoopSettings\TraceLoopSettings;
 use PHPUnit\Framework\TestCase;
 
 class WorkerPoolTest extends TestCase
