+++ tests/Inspector/Settings/TargetProcessSettings/TargetProcessSettingsTest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace PhpProfiler\Inspector\Settings;
+namespace PhpProfiler\Inspector\Settings\TargetProcessSettings;
 
 use Mockery;
 use PHPUnit\Framework\TestCase;
